[BugFix] Fix a bug (works correctly in practice) in serialize method of CELONIS_BUILD_KMEANS_MODEL (#1766)
## Why I'm doing:

num_clusters_ is int64_t. In the original implementation of serialize,
dst += sizeof(double); is used. The code works correctly in practice,
because sizeof(int64_t) == sizeof(double) == 8 on 64-bit system. However
it is still technically a bug from a code maintenance perspective.

## What I'm doing:

This PR fixes this bug.

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5

+++ be/src/exprs/celonis/agg/kmeans.h
@@ -78,7 +78,7 @@ struct CelonisKMeansModelAggregateState {
         memcpy(dst, &inconsistent_dimension_, sizeof(uint8_t));
         dst += sizeof(uint8_t);
         memcpy(dst, &num_clusters_, sizeof(int64_t));
-        dst += sizeof(double);
+        dst += sizeof(int64_t);
         memcpy(dst, &seed_, sizeof(int));
         dst += sizeof(int);
         memcpy(dst, &num_features_, sizeof(size_t));
